feat: Refactor currency data model and implement currency conversion
This commit refactors the currency data model, implements currency conversion functionality, and updates the UI to support these changes.

**Key changes:**

-   **Domain Layer:**
    -   Created `CurrencyInfo` data class to hold static currency information (title, string resource ID, type, ID).
    -   Updated `Currency` data class to include `CurrencyInfo` and use `Int` for `currentPrice`.
    -   Created `Calculator` interface for currency conversion logic.
-   **Data Layer:**
    -   Implemented `CurrencyCalculator` to calculate currency exchange rates using `CurrenciesRepository`.
    -   Updated `CurrencyMapper` to use `CurrencyInfo` and `Int` for `currentPrice`.
    -   Updated `CurrencyEntity` to use `Int` for `currentPrice`.
-   **Presentation Layer:**
    -   Renamed `MainAction` to `CurrencyAction` and added actions for currency conversion.
    -   Renamed `MainState` to `CurrencyState` and added properties for currency conversion UI.
    -   Created `CurrencyScreen` to display currency data and handle user interactions.
    -   Created `ConvertCurrencyModalBottomSheet` and `SelectCurrencyModal` for currency conversion UI.
    -   Updated `CurrencyItem` to use `CurrencyInfo`.
    -   Implemented a snackbar for displaying error messages.
-   **Dependency Injection:**
    -   Registered `Calculator`, `Navigator`, and `CurrencyViewModel` in `appModule`.
-   **Other Changes:**
    -   Updated `CurrencyDao` to order currencies by ID.
    -   Added a visual transformation for currency input fields.
    -   Added an error color to the theme.
    -   Implemented a visual transformation for currency input fields.
    -   Added an error color to the theme.
    -   Updated `PullToRefreshBox` to use `isRefreshing` for refresh indicator visibility.

+++ app/src/main/java/ir/moodz/sarafkoochooloo/data/local/entity/CurrencyEntity.kt
@@ -9,7 +9,7 @@ data class CurrencyEntity(
     val id: Int,
     @PrimaryKey(autoGenerate = false)
     val title: String,
-    val currentPrice: Long,
+    val currentPrice: Int,
     val updatedTime: String,
     val type: CurrencyType
 )
